(svn r63) Fix: [ 1009385 ] Too many save games prevented loading

+++ window.h
@@ -162,7 +162,7 @@ enum {
 #define WP(ptr,str) (*(str*)(ptr)->custom)
 
 typedef struct {
-	byte count, cap, pos;
+	uint16 count, cap, pos;
 } Scrollbar;
 
 struct Window {
